[session 6.14.3] ** Org-default UX polish, resolver audit, and test policy alignment
